+++ scriptabit/plugins/health_effects.py
@@ -1,0 +1,12 @@
+""" Poisoning and health regeneration plugin.
+"""
+
+from scriptabit.plugin_interfaces import IOfficialPlugin
+
+
+class HealthEffects(IOfficialPlugin):
+    """Official scriptabit plugin that implements a poisoning/health
+    regeneration scenario based on player performance on dailies.
+    """
+
+    pass